requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.4.9
aiohttp==3.9.5
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
redis==5.0.1
pydantic==2.5.0